+++ package/kernel/sx1276-spidevs/src/RegionEU433.c
@@ -511,7 +511,7 @@ bool RegionEU433AdrNext( AdrNextParams_t* adrNext, int8_t* drOut, int8_t* txPowO
 
 void RegionEU433ComputeRxWindowParameters( int8_t datarate, uint8_t minRxSymbols, uint32_t rxError, RxConfigParams_t *rxConfigParams )
 {
-    double tSymbol = 0.0;
+    long tSymbol = 0.0;
 
     // Get the datarate, perform a boundary check
     rxConfigParams->Datarate = MIN( datarate, EU433_RX_MAX_DATARATE );
